refactor: 폰북 뷰 컨트롤러에서 폰북 뷰 분리
