+++ phonebook/Controller/PhoneBookViewController.swift
@@ -9,84 +9,21 @@ import UIKit
 import SnapKit
 
 class PhoneBookViewController: UIViewController {
-    // MARK: - UI Components
-    private let profileImageView: UIImageView = {
-        let imageView = UIImageView()
-        imageView.layer.cornerRadius = 80
-        imageView.layer.borderWidth = 1
-        imageView.layer.borderColor = UIColor.lightGray.cgColor
-        imageView.clipsToBounds = true
-        imageView.backgroundColor = .systemGray6
-        return imageView
-    }()
-    
-    private let randomImageButton: UIButton = {
-        let button = UIButton(type: .system)
-        button.setTitle("랜덤 이미지 생성", for: .normal)
-        button.titleLabel?.font = UIFont.systemFont(ofSize: 14, weight: .medium)
-        button.setTitleColor(.gray, for: .normal)
-        return button
-    }()
-    
-    private let nameTextView: UITextView = {
-        let textView = UITextView()
-        textView.font = UIFont.systemFont(ofSize: 16)
-        textView.layer.borderWidth = 1
-        textView.layer.borderColor = UIColor.lightGray.cgColor
-        textView.layer.cornerRadius = 8
-        textView.backgroundColor = .white
-        return textView
-    }()
-    
-    private let phoneTextView: UITextView = {
-        let textView = UITextView()
-        textView.font = UIFont.systemFont(ofSize: 16)
-        textView.layer.borderWidth = 1
-        textView.layer.borderColor = UIColor.lightGray.cgColor
-        textView.layer.cornerRadius = 8
-        textView.backgroundColor = .white
-        return textView
-    }()
+    let phoneBookView = PhoneBookView()
     
     // MARK: - Life Cycle
+    override func loadView() {
+        view = phoneBookView
+    }
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         view.backgroundColor = .white
-        setupUI()
         setupNavigationBar()
+        phoneBookView.randomImageButton.addTarget(self, action: #selector(randomImageButtonTapped), for: .touchUpInside)
     }
     
     // MARK: - Setup Methods
-    private func setupUI() {
-        view.addSubview(profileImageView)
-        view.addSubview(randomImageButton)
-        view.addSubview(nameTextView)
-        view.addSubview(phoneTextView)
-        
-        profileImageView.snp.makeConstraints { make in
-            make.top.equalTo(view.safeAreaLayoutGuide.snp.top).offset(16)
-            make.centerX.equalToSuperview()
-            make.size.equalTo(160)
-        }
-        
-        randomImageButton.snp.makeConstraints { make in
-            make.top.equalTo(profileImageView.snp.bottom).offset(8)
-            make.centerX.equalToSuperview()
-        }
-        
-        nameTextView.snp.makeConstraints { make in
-            make.top.equalTo(randomImageButton.snp.bottom).offset(24)
-            make.leading.trailing.equalToSuperview().inset(16)
-            make.height.equalTo(40)
-        }
-        
-        phoneTextView.snp.makeConstraints { make in
-            make.top.equalTo(nameTextView.snp.bottom).offset(8)
-            make.leading.trailing.equalToSuperview().inset(16)
-            make.height.equalTo(40)
-        }
-    }
-    
     private func setupNavigationBar() {
         title = "연락처 추가"
         navigationItem.rightBarButtonItem = UIBarButtonItem(
@@ -97,6 +34,12 @@ class PhoneBookViewController: UIViewController {
     private func saveButtonTapped() {
         print("save!!")
         // TODO : save function
+    }
+    
+    @objc
+    private func randomImageButtonTapped() {
+        print("random")
+        // TODO : random image function
     }
 }
